Change delete forms for Appointments & Relaunches in Application view

+++ src/JG/AdminBundle/Controller/Account/AccountRelaunchController.php
@@ -101,19 +101,29 @@ class AccountRelaunchController extends Controller
     /**
      * Deletes a relaunch entity.
      *
-     * @Route("/{id}/delete", name="relaunch_delete_link")
+     * @Route("/{id}/delete", name="relaunch_delete")
+     * @Method("DELETE")
      */
-    public function deleteRelaunchAction(Request $request, Relaunch $relaunch)
+    public function deleteAction(Request $request, Relaunch $relaunch)
     {
-        $application = $relaunch->getApplication();
+        $form = $this->createDeleteForm($relaunch);
+        $form->handleRequest($request);
 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($relaunch);
-        $em->flush($relaunch);
+        $user = $this->getUser();
 
-        $request->getSession()->getFlashBag()->add('success', 'Relance supprimée avec succès !');
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $relaunch->getApplication()->removeRelaunch($relaunch);
+            $em->remove($relaunch);
+            $em->flush($relaunch);
+            $request->getSession()->getFlashBag()->add('success', 'Relance supprimée avec succès !');
+            return $this->redirectToRoute('application_show', array('id' => $relaunch->getApplication()->getId()));
+        }
 
-        return $this->redirectToRoute('application_show', array('id' => $application->getId()));
+        return $this->render('JGAdminBundle:Account:appointment/delete.html.twig', array(
+            'relaunch'   => $relaunch,
+            'delete_form'   => $form->createView(),
+        ));
     }
 
     /**
@@ -123,12 +133,12 @@ class AccountRelaunchController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    /*private function createDeleteForm(Relaunch $relaunch)
+    private function createDeleteForm(Relaunch $relaunch)
     {
         return $this->createFormBuilder()
             ->setAction($this->generateUrl('relaunch_delete', array('id' => $relaunch->getId())))
             ->setMethod('DELETE')
             ->getForm()
         ;
-    }*/
+    }
 }
